increasing compatibility with older Matlabs
Testing on R2012a…

+++ mfiles/@spinw/genmagstr.m
@@ -206,7 +206,7 @@ inpForm.soft   = [inpForm.soft   {true      true  false  false     false  }];
 
 param = sw_readparam(inpForm, varargin{:});
 
-if prod(param.nExt) == 0
+if prod(double(param.nExt)) == 0
     error('spinw:genmagstr:WrongInput','''nExt'' has to be larger than 0!');
 end
